Use shgo instead of minimize
Global optimization is less likely to get stuck when the function
has discrete plateaus, like in the case of simulated data (aliasing).
At the same time, shgo is fast like gradient-based methods for this use
case.

The conditioning is improved as well, otherwise the overfocus could
overshoot and the rotation was less effective.

Tested on both simulated and real data.