Add a CI job that tests against pinned requirements
As these requirements are the ones that end up in the docker container
image, we should make updating the requirements testable. The
installation tries to mimick what's done in the container build script
as far as possible, but as it needs to install the `[test]` extra, it
can't use `--no-deps` and thus isn't a 100% match.